chapter16/requirements.txt
# CogVLM2
xformers==0.0.23.post1
torch==2.1.2
torchvision==0.16.2
transformers==4.40
huggingface-hub==0.23.0
pillow==10.4.0
chainlit==1.1.306
pydantic==2.7.1
timm==0.9.16
openai==1.30.1
loguru==0.7.2
pydantic==2.7.1
einops==0.8.0
sse-starlette==2.1.0
bitsandbytes==0.43.1
accelerate==0.32.1
# Langchain
langchain_community==0.2.7
sentence-transformers==3.0.1
chromadb==0.5.3
# Webui
gradio==4.37.2